requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Cloud
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.0

# LangGraph
langgraph==0.0.20

# Image/Video Processing
opencv-python==4.8.1.78
Pillow==10.1.0
ffmpeg-python==0.2.0

# CLIP
torch==2.2.0
torchvision==0.17.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.0

# Logging
loguru==0.7.2

